extract status update to custom hook
